+++ pybop/optimisers/pints_optimisers.py
@@ -1,6 +1,15 @@
-import pints
+from warnings import warn
 
-from pybop import BasePintsOptimiser
+from pints import CMAES as PintsCMAES
+from pints import PSO as PintsPSO
+from pints import SNES as PintsSNES
+from pints import XNES as PintsXNES
+from pints import Adam as PintsAdam
+from pints import GradientDescent as PintsGradientDescent
+from pints import IRPropMin as PintsIRPropMin
+from pints import NelderMead as PintsNelderMead
+
+from pybop import AdamWImpl, BasePintsOptimiser
 
 
 class GradientDescent(BasePintsOptimiser):
@@ -28,7 +37,7 @@ class GradientDescent(BasePintsOptimiser):
     def __init__(self, cost, **optimiser_kwargs):
         if "sigma0" not in optimiser_kwargs.keys():
             optimiser_kwargs["sigma0"] = 0.02  # set default
-        super().__init__(cost, pints.GradientDescent, **optimiser_kwargs)
+        super().__init__(cost, PintsGradientDescent, **optimiser_kwargs)
 
 
 class Adam(BasePintsOptimiser):
@@ -53,8 +62,41 @@ class Adam(BasePintsOptimiser):
     pints.Adam : The PINTS implementation this class is based on.
     """
 
+    warn(
+        "Adam is deprecated and will be removed in a future release. Please use AdamW instead.",
+        DeprecationWarning,
+        stacklevel=2,
+    )
+
     def __init__(self, cost, **optimiser_kwargs):
-        super().__init__(cost, pints.Adam, **optimiser_kwargs)
+        super().__init__(cost, PintsAdam, **optimiser_kwargs)
+
+
+class AdamW(BasePintsOptimiser):
+    """
+    Implements the AdamW optimisation algorithm in PyBOP.
+
+    This class extends the AdamW optimiser, which is a variant of the Adam
+    optimiser that incorporates weight decay. AdamW is designed to be more
+    robust and stable for training deep neural networks, particularly when
+    using larger learning rates.
+
+    Parameters
+    ----------
+    **optimiser_kwargs : optional
+        Valid PyBOP option keys and their values, for example:
+        x0 : array_like
+            Initial position from which optimisation will start.
+        sigma0 : float
+            Initial step size.
+
+    See Also
+    --------
+    pybop.AdamWImpl : The PyBOP implementation this class is based on.
+    """
+
+    def __init__(self, cost, **optimiser_kwargs):
+        super().__init__(cost, AdamWImpl, **optimiser_kwargs)
 
 
 class IRPropMin(BasePintsOptimiser):
@@ -83,7 +125,7 @@ class IRPropMin(BasePintsOptimiser):
     """
 
     def __init__(self, cost, **optimiser_kwargs):
-        super().__init__(cost, pints.IRPropMin, **optimiser_kwargs)
+        super().__init__(cost, PintsIRPropMin, **optimiser_kwargs)
 
 
 class PSO(BasePintsOptimiser):
@@ -112,7 +154,7 @@ class PSO(BasePintsOptimiser):
     """
 
     def __init__(self, cost, **optimiser_kwargs):
-        super().__init__(cost, pints.PSO, **optimiser_kwargs)
+        super().__init__(cost, PintsPSO, **optimiser_kwargs)
 
 
 class SNES(BasePintsOptimiser):
@@ -141,7 +183,7 @@ class SNES(BasePintsOptimiser):
     """
 
     def __init__(self, cost, **optimiser_kwargs):
-        super().__init__(cost, pints.SNES, **optimiser_kwargs)
+        super().__init__(cost, PintsSNES, **optimiser_kwargs)
 
 
 class XNES(BasePintsOptimiser):
@@ -170,7 +212,7 @@ class XNES(BasePintsOptimiser):
     """
 
     def __init__(self, cost, **optimiser_kwargs):
-        super().__init__(cost, pints.XNES, **optimiser_kwargs)
+        super().__init__(cost, PintsXNES, **optimiser_kwargs)
 
 
 class NelderMead(BasePintsOptimiser):
@@ -197,7 +239,7 @@ class NelderMead(BasePintsOptimiser):
     """
 
     def __init__(self, cost, **optimiser_kwargs):
-        super().__init__(cost, pints.NelderMead, **optimiser_kwargs)
+        super().__init__(cost, PintsNelderMead, **optimiser_kwargs)
 
 
 class CMAES(BasePintsOptimiser):
@@ -232,4 +274,4 @@ class CMAES(BasePintsOptimiser):
                 "CMAES requires optimisation of >= 2 parameters at once. "
                 + "Please choose another optimiser."
             )
-        super().__init__(cost, pints.CMAES, **optimiser_kwargs)
+        super().__init__(cost, PintsCMAES, **optimiser_kwargs)